Remove Checkin model from JS layer
