+++ src/MyHealth/SymptomLogContext.tsx
@@ -8,25 +8,18 @@ import React, {
 
 import {
   SymptomLogEntry,
-  CheckIn,
-  CheckInStatus,
   DayLogData,
   combineSymptomAndCheckInLogs,
   Symptom,
 } from "./symptoms"
 import {
   getLogEntries,
-  getCheckIns,
-  addCheckIn,
   createLogEntry,
   modifyLogEntry,
   deleteLogEntry as removeLogEntry,
-  deleteAllCheckIns as deleteCheckIns,
   deleteAllSymptomLogs as deleteLogs,
-  deleteStaleCheckIns,
   deleteStaleSymptomLogs,
 } from "./nativeModule"
-import { isToday } from "../utils/dateTime"
 import {
   failureResponse,
   OperationResponse,
@@ -38,9 +31,6 @@ export type SymptomLogState = {
   addLogEntry: (symptoms: Symptom[]) => Promise<OperationResponse>
   updateLogEntry: (entry: SymptomLogEntry) => Promise<OperationResponse>
   deleteLogEntry: (symptomLogEntryId: string) => Promise<OperationResponse>
-  todaysCheckIn: CheckIn
-  addTodaysCheckIn: (status: CheckInStatus) => Promise<OperationResponse>
-  deleteAllCheckIns: () => Promise<OperationResponse>
   deleteAllLogEntries: () => Promise<OperationResponse>
 }
 
@@ -55,13 +45,6 @@ const initialState: SymptomLogState = {
   deleteLogEntry: (_symptomLogEntryId: string) => {
     return Promise.resolve(SUCCESS_RESPONSE)
   },
-  todaysCheckIn: { date: Date.now(), status: CheckInStatus.NotCheckedIn },
-  addTodaysCheckIn: (_status: CheckInStatus) => {
-    return Promise.resolve(SUCCESS_RESPONSE)
-  },
-  deleteAllCheckIns: () => {
-    return Promise.resolve(SUCCESS_RESPONSE)
-  },
   deleteAllLogEntries: () => {
     return Promise.resolve(SUCCESS_RESPONSE)
   },
@@ -72,17 +55,6 @@ export const SymptomLogContext = createContext<SymptomLogState>(initialState)
 export const SymptomLogProvider: FunctionComponent = ({ children }) => {
   const [dailyLogData, setDailyLogData] = useState<DayLogData[]>([])
   const [logEntries, setLogEntries] = useState<SymptomLogEntry[]>([])
-  const [checkIns, setCheckIns] = useState<CheckIn[]>([])
-  const [todaysCheckIn, setTodaysCheckIn] = useState<CheckIn>(
-    initialState.todaysCheckIn,
-  )
-
-  const detectTodaysCheckIn = (rawCheckIns: CheckIn[]) => {
-    const checkInAddedToday = rawCheckIns.find(({ date }) => {
-      return isToday(date)
-    })
-    checkInAddedToday && setTodaysCheckIn(checkInAddedToday)
-  }
 
   const fetchLogEntries = async () => {
     const entries = await getLogEntries()
@@ -90,22 +62,17 @@ export const SymptomLogProvider: FunctionComponent = ({ children }) => {
   }
 
   const cleanupStaleData = async () => {
-    await deleteStaleCheckIns()
     await deleteStaleSymptomLogs()
   }
 
   useEffect(() => {
     cleanupStaleData()
     fetchLogEntries()
-    getCheckIns().then((checkIns) => {
-      setCheckIns(checkIns)
-      detectTodaysCheckIn(checkIns)
-    })
   }, [])
 
   useEffect(() => {
-    setDailyLogData(combineSymptomAndCheckInLogs(logEntries, checkIns))
-  }, [checkIns, logEntries])
+    setDailyLogData(combineSymptomAndCheckInLogs(logEntries))
+  }, [logEntries])
 
   const addLogEntry = async (symptoms: Symptom[]) => {
     try {
@@ -141,30 +108,6 @@ export const SymptomLogProvider: FunctionComponent = ({ children }) => {
     }
   }
 
-  const addTodaysCheckIn = async (status: CheckInStatus) => {
-    try {
-      const newCheckIn = { date: Date.now(), status }
-      await addCheckIn(newCheckIn)
-      setTodaysCheckIn(newCheckIn)
-      const newCheckIns = await getCheckIns()
-      setCheckIns(newCheckIns)
-      return SUCCESS_RESPONSE
-    } catch (e) {
-      return failureResponse(e.message)
-    }
-  }
-
-  const deleteAllCheckIns = async () => {
-    try {
-      await deleteCheckIns()
-      await getCheckIns()
-      setTodaysCheckIn(initialState.todaysCheckIn)
-      return SUCCESS_RESPONSE
-    } catch (e) {
-      return failureResponse(e.message)
-    }
-  }
-
   const deleteAllLogEntries = async () => {
     try {
       await deleteLogs()
@@ -178,13 +121,10 @@ export const SymptomLogProvider: FunctionComponent = ({ children }) => {
   return (
     <SymptomLogContext.Provider
       value={{
-        addTodaysCheckIn,
-        todaysCheckIn,
         dailyLogData,
         addLogEntry,
         updateLogEntry,
         deleteLogEntry,
-        deleteAllCheckIns,
         deleteAllLogEntries,
       }}
     >
