+++ src/MyHealth/index.tsx
@@ -1,11 +1,9 @@
-import React, { FunctionComponent, useState } from "react"
+import React, { FunctionComponent } from "react"
 import { View, StyleSheet } from "react-native"
-import SegmentedControl from "@react-native-community/segmented-control"
 import { useTranslation } from "react-i18next"
 
 import { useStatusBarEffect } from "../navigation"
 import { GlobalText, StatusBar } from "../components"
-import Today from "./Today"
 import OverTime from "./OverTime"
 
 import { Typography, Colors, Spacing } from "../styles"
@@ -13,12 +11,6 @@ import { Typography, Colors, Spacing } from "../styles"
 const MyHealthScreen: FunctionComponent = () => {
   useStatusBarEffect("dark-content", Colors.secondary10)
   const { t } = useTranslation()
-  const [selectedIndex, setSelectedIndex] = useState<number>(0)
-
-  const todayStyle = selectedIndex === 0 ? style.innerContainer : style.hidden
-
-  const overTimeStyle =
-    selectedIndex === 1 ? style.innerContainer : style.hidden
 
   return (
     <>
@@ -28,22 +20,8 @@ const MyHealthScreen: FunctionComponent = () => {
           <GlobalText style={style.headerText}>
             {t("symptom_checker.my_health")}
           </GlobalText>
-          <View style={style.segmentedControlContainer}>
-            <SegmentedControl
-              values={[t("my_health.today"), t("my_health.over_time")]}
-              selectedIndex={selectedIndex}
-              onChange={(event) => {
-                setSelectedIndex(event.nativeEvent.selectedSegmentIndex)
-              }}
-              fontStyle={style.segmentedControlText}
-              activeFontStyle={style.activeSegmentedControlText}
-            />
-          </View>
         </View>
-        <View style={todayStyle}>
-          <Today />
-        </View>
-        <View style={overTimeStyle}>
+        <View style={style.innerContainer}>
           <OverTime />
         </View>
       </View>
