+++ src/MyHealth/nativeModule.ts
@@ -1,29 +1,9 @@
 import { NativeModules } from "react-native"
 
 import {
-  CheckIn,
   SymptomLogEntry,
   SymptomLogEntryAttributes,
 } from "../MyHealth/symptoms"
-
-// Check In Module
-const checkInModule = NativeModules.CheckInModule
-
-export const getCheckIns = async (): Promise<CheckIn[]> => {
-  return checkInModule.getCheckIns()
-}
-
-export const addCheckIn = (checkIn: CheckIn): Promise<void> => {
-  return checkInModule.addCheckIn(checkIn)
-}
-
-export const deleteAllCheckIns = async (): Promise<"success"> => {
-  return checkInModule.deleteCheckins()
-}
-
-export const deleteStaleCheckIns = async (): Promise<"success"> => {
-  return checkInModule.deleteStaleCheckIns()
-}
 
 // Symptom Log Entry Module
 const symptomLogEntryModule = NativeModules.SymptomLogEntryModule
