+++ src/Settings/ENDebugMenu.tsx
@@ -27,7 +27,7 @@ const ENDebugMenu: FunctionComponent<ENDebugMenuProps> = ({ navigation }) => {
   useStatusBarEffect("light-content", Colors.headerBackground)
   const [loading, setLoading] = useState(false)
   const { resetOnboarding } = useOnboardingContext()
-  const { deleteAllCheckIns, deleteAllLogEntries } = useSymptomLogContext()
+  const { deleteAllLogEntries } = useSymptomLogContext()
 
   useEffect(() => {
     const handleBackPress = () => {
@@ -135,13 +135,6 @@ const ENDebugMenu: FunctionComponent<ENDebugMenuProps> = ({ navigation }) => {
             <DebugMenuListItem
               label="Restart Onboarding"
               onPress={handleOnPressRestartOnboarding}
-            />
-            <DebugMenuListItem
-              label="Delete All CheckIns"
-              onPress={handleOnPressSimulationButton(async () => {
-                const result = await deleteAllCheckIns()
-                return Promise.resolve(result.kind)
-              })}
             />
             <DebugMenuListItem
               label="Delete All Symptom Logs"
